[WIP] Factor BoundedQuantityValue out of QuantityValue
