+++ src/DataValues/QuantityValue.php
@@ -5,8 +5,10 @@ namespace DataValues;
 use InvalidArgumentException;
 
 /**
- * Class representing a quantity with associated unit and uncertainty interval.
+ * Class representing a quantity with associated unit.
  * The amount is stored as a @see DecimalValue object.
+ *
+ * For quantities with a known uncertainty interval, see BoundedQuantityValue.
  *
  * For simple numeric amounts use @see NumberValue.
  *
@@ -22,28 +24,14 @@ class QuantityValue extends DataValueObject {
 	 *
 	 * @var DecimalValue
 	 */
-	private $amount;
+	protected $amount;
 
 	/**
 	 * The quantity's unit identifier (use "1" for unitless quantities).
 	 *
 	 * @var string
 	 */
-	private $unit;
-
-	/**
-	 * The quantity's upper bound
-	 *
-	 * @var DecimalValue
-	 */
-	private $upperBound;
-
-	/**
-	 * The quantity's lower bound
-	 *
-	 * @var DecimalValue
-	 */
-	private $lowerBound;
+	protected $unit;
 
 	/**
 	 * Constructs a new QuantityValue object, representing the given value.
@@ -52,20 +40,10 @@ class QuantityValue extends DataValueObject {
 	 *
 	 * @param DecimalValue $amount
 	 * @param string $unit A unit identifier. Must not be empty, use "1" for unit-less quantities.
-	 * @param DecimalValue $upperBound The upper bound of the quantity, inclusive.
-	 * @param DecimalValue $lowerBound The lower bound of the quantity, inclusive.
 	 *
 	 * @throws IllegalValueException
 	 */
-	public function __construct( DecimalValue $amount, $unit, DecimalValue $upperBound, DecimalValue $lowerBound ) {
-		if ( $lowerBound->compare( $amount ) > 0 ) {
-			throw new IllegalValueException( '$lowerBound ' . $lowerBound->getValue() . ' must be <= $amount ' . $amount->getValue() );
-		}
-
-		if ( $upperBound->compare( $amount ) < 0 ) {
-			throw new IllegalValueException( '$upperBound ' . $upperBound->getValue() . ' must be >= $amount ' . $amount->getValue() );
-		}
-
+	public function __construct( DecimalValue $amount, $unit ) {
 		if ( !is_string( $unit ) ) {
 			throw new IllegalValueException( '$unit needs to be a string, not ' . gettype( $unit ) );
 		}
@@ -76,14 +54,10 @@ class QuantityValue extends DataValueObject {
 
 		$this->amount = $amount;
 		$this->unit = $unit;
-		$this->upperBound = $upperBound;
-		$this->lowerBound = $lowerBound;
 	}
 
 	/**
 	 * Returns a QuantityValue representing the given amount.
-	 * If no upper or lower bound is given, the amount is assumed to be absolutely exact,
-	 * that is, the amount itself will be used as the upper and lower bound.
 	 *
 	 * This is a convenience wrapper around the constructor that accepts native values
 	 * instead of DecimalValue objects.
@@ -93,20 +67,16 @@ class QuantityValue extends DataValueObject {
 	 *
 	 * @since 0.1
 	 *
-	 * @param string|int|float|DecimalValue $amount
+	 * @param string|int|float|DecimalValue $number
 	 * @param string $unit A unit identifier. Must not be empty, use "1" for unit-less quantities.
-	 * @param string|int|float|DecimalValue|null $upperBound
-	 * @param string|int|float|DecimalValue|null $lowerBound
 	 *
 	 * @return self
 	 * @throws IllegalValueException
 	 */
-	public static function newFromNumber( $amount, $unit = '1', $upperBound = null, $lowerBound = null ) {
-		$amount = self::asDecimalValue( 'amount', $amount );
-		$upperBound = self::asDecimalValue( 'upperBound', $upperBound, $amount );
-		$lowerBound = self::asDecimalValue( 'lowerBound', $lowerBound, $amount );
+	public static function newFromNumber( $number, $unit = '1' ) {
+		$number = self::asDecimalValue( 'amount', $number );
 
-		return new self( $amount, $unit, $upperBound, $lowerBound );
+		return new self( $number, $unit );
 	}
 
 	/**
@@ -114,15 +84,15 @@ class QuantityValue extends DataValueObject {
 	 *
 	 * @deprecated since 0.1, use newFromNumber instead
 	 *
-	 * @param string|int|float|DecimalValue $amount
+	 * @param string|int|float|DecimalValue $number
 	 * @param string $unit
 	 * @param string|int|float|DecimalValue|null $upperBound
 	 * @param string|int|float|DecimalValue|null $lowerBound
 	 *
 	 * @return self
 	 */
-	public static function newFromDecimal( $amount, $unit = '1', $upperBound = null, $lowerBound = null ) {
-		return self::newFromNumber( $amount, $unit, $upperBound, $lowerBound );
+	public static function newFromDecimal( $number, $unit = '1' ) {
+		return self::newFromNumber( $number, $unit );
 	}
 
 	/**
@@ -139,7 +109,7 @@ class QuantityValue extends DataValueObject {
 	 * @throws InvalidArgumentException
 	 * @return DecimalValue
 	 */
-	private static function asDecimalValue( $name, $number, DecimalValue $default = null ) {
+	protected static function asDecimalValue( $name, $number, DecimalValue $default = null ) {
 		if ( !is_string( $name ) ) {
 			throw new InvalidArgumentException( '$name must be a string' );
 		}
@@ -171,12 +141,7 @@ class QuantityValue extends DataValueObject {
 	 * @return string
 	 */
 	public function serialize() {
-		return serialize( array(
-			$this->amount,
-			$this->unit,
-			$this->upperBound,
-			$this->lowerBound,
-		) );
+		return serialize( array_values( $this->getArrayValue() ) );
 	}
 
 	/**
@@ -187,8 +152,8 @@ class QuantityValue extends DataValueObject {
 	 * @param string $data
 	 */
 	public function unserialize( $data ) {
-		list( $amount, $unit, $upperBound, $lowerBound ) = unserialize( $data );
-		$this->__construct( $amount, $unit, $upperBound, $lowerBound );
+		list( $amount, $unit ) = unserialize( $data );
+		$this->__construct( $amount, $unit );
 	}
 
 	/**
@@ -234,128 +199,6 @@ class QuantityValue extends DataValueObject {
 	 */
 	public function getAmount() {
 		return $this->amount;
-	}
-
-	/**
-	 * Returns this quantity's upper bound.
-	 *
-	 * @since 0.1
-	 *
-	 * @return DecimalValue
-	 */
-	public function getUpperBound() {
-		return $this->upperBound;
-	}
-
-	/**
-	 * Returns this quantity's lower bound.
-	 *
-	 * @since 0.1
-	 *
-	 * @return DecimalValue
-	 */
-	public function getLowerBound() {
-		return $this->lowerBound;
-	}
-
-	/**
-	 * Returns the size of the uncertainty interval.
-	 * This can roughly be interpreted as "amount +/- uncertainty/2".
-	 *
-	 * The exact interpretation of the uncertainty interval is left to the concrete application or
-	 * data point. For example, the uncertainty interval may be defined to be that part of a
-	 * normal distribution that is required to cover the 95th percentile.
-	 *
-	 * @since 0.1
-	 *
-	 * @return float
-	 */
-	public function getUncertainty() {
-		return $this->upperBound->getValueFloat() - $this->lowerBound->getValueFloat();
-	}
-
-	/**
-	 * Returns a DecimalValue representing the symmetrical offset to be applied
-	 * to the raw amount for a rough representation of the uncertainty interval,
-	 * as in "amount +/- offset".
-	 *
-	 * The offset is calculated as max( amount - lowerBound, upperBound - amount ).
-	 *
-	 * @since 0.1
-	 *
-	 * @return DecimalValue
-	 */
-	public function getUncertaintyMargin() {
-		$math = new DecimalMath();
-
-		$lowerMargin = $math->sum( $this->amount, $this->lowerBound->computeComplement() );
-		$upperMargin = $math->sum( $this->upperBound, $this->amount->computeComplement() );
-
-		$margin = $math->max( $lowerMargin, $upperMargin );
-		return $margin;
-	}
-
-	/**
-	 * Returns the order of magnitude of the uncertainty as the exponent of
-	 * last significant digit in the amount-string. The value returned by this
-	 * is suitable for use with @see DecimalMath::roundToExponent().
-	 *
-	 * @example: if two digits after the decimal point are significant, this
-	 * returns -2.
-	 *
-	 * @example: if the last two digits before the decimal point are insignificant,
-	 * this returns 2.
-	 *
-	 * Note that this calculation assumes a symmetric uncertainty interval,
-	 * and can be misleading.
-	 *
-	 * @since 0.1
-	 *
-	 * @return int
-	 */
-	public function getOrderOfUncertainty() {
-		// the desired precision is given by the distance between the amount and
-		// whatever is closer, the upper or lower bound.
-		//TODO: use DecimalMath to avoid floating point errors!
-		$amount = $this->amount->getValueFloat();
-		$upperBound = $this->upperBound->getValueFloat();
-		$lowerBound = $this->lowerBound->getValueFloat();
-		$precision = min( $amount - $lowerBound, $upperBound - $amount );
-
-		if ( $precision === 0.0 ) {
-			// If there is no uncertainty, the order of uncertainty is a bit more than what we have digits for.
-			return -strlen( $this->amount->getFractionalPart() );
-		}
-
-		// e.g. +/- 200 -> 2; +/- 0.02 -> -2
-		// note: we really want floor( log10( $precision ) ), but have to account for
-		// small errors made in the floating point operations above.
-		// @todo: use bcmath (via DecimalMath) to avoid this if possible
-		$orderOfUncertainty = floor( log10( $precision + 0.0000000005 ) );
-
-		return (int)$orderOfUncertainty;
-	}
-
-	/**
-	 * Returns the number of significant figures in the amount-string,
-	 * counting the decimal point, but not counting the leading sign.
-	 *
-	 * Note that this calculation assumes a symmetric uncertainty interval, and can be misleading
-	 *
-	 * @since 0.1
-	 *
-	 * @return int
-	 */
-	public function getSignificantFigures() {
-		$math = new DecimalMath();
-
-		// $orderOfUncertainty is +/- 200 -> 2; +/- 0.02 -> -2
-		$orderOfUncertainty = $this->getOrderOfUncertainty();
-
-		// the number of digits (without the sign) is the same as the position (with the sign).
-		$significantDigits = $math->getPositionForExponent( $orderOfUncertainty, $this->amount );
-
-		return $significantDigits;
 	}
 
 	/**
@@ -425,31 +268,12 @@ class QuantityValue extends DataValueObject {
 		$args[0] = $this->amount;
 		$amount = call_user_func_array( $transformation, $args );
 
-		$args[0] = $this->upperBound;
-		$upperBound = call_user_func_array( $transformation, $args );
-
-		$args[0] = $this->lowerBound;
-		$lowerBound = call_user_func_array( $transformation, $args );
-
 		// use a preliminary QuantityValue to determine the significant number of digits
-		$transformed = new self( $amount, $newUnit, $upperBound, $lowerBound );
-		$roundingExponent = $transformed->getOrderOfUncertainty();
-
-		// apply rounding to the significant digits
-		$math = new DecimalMath();
-
-		$amount = $math->roundToExponent( $amount, $roundingExponent );
-		$upperBound = $math->roundToExponent( $upperBound, $roundingExponent );
-		$lowerBound = $math->roundToExponent( $lowerBound, $roundingExponent );
-
-		return new self( $amount, $newUnit, $upperBound, $lowerBound );
+		return new self( $amount, $newUnit );
 	}
 
 	public function __toString() {
 		return $this->amount->getValue()
-			. '[' . $this->lowerBound->getValue()
-			. '..' . $this->upperBound->getValue()
-			. ']'
 			. ( $this->unit === '1' ? '' : $this->unit );
 	}
 
@@ -464,8 +288,6 @@ class QuantityValue extends DataValueObject {
 		return array(
 			'amount' => $this->amount->getArrayValue(),
 			'unit' => $this->unit,
-			'upperBound' => $this->upperBound->getArrayValue(),
-			'lowerBound' => $this->lowerBound->getArrayValue(),
 		);
 	}
 
@@ -481,13 +303,11 @@ class QuantityValue extends DataValueObject {
 	 * @throws IllegalValueException
 	 */
 	public static function newFromArray( $data ) {
-		self::requireArrayFields( $data, array( 'amount', 'unit', 'upperBound', 'lowerBound' ) );
+		self::requireArrayFields( $data, array( 'amount', 'unit' ) );
 
 		return new static(
 			DecimalValue::newFromArray( $data['amount'] ),
-			$data['unit'],
-			DecimalValue::newFromArray( $data['upperBound'] ),
-			DecimalValue::newFromArray( $data['lowerBound'] )
+			$data['unit']
 		);
 	}
